refactor: ProductAdapterEventHandler 구현체를 ViewModel로 이동
